+++ src/components/layout/sideMenu/parts/MenuItemWithSubmenu.tsx
@@ -172,7 +172,7 @@ export const MenuItemWithSubmenu = ({
 
   const showTooltip = isCollapsed && hasEnteredSinceCollapse && !isDropdownOpen;
 
-  const sharedClassName = `flex relative rounded-md items-center py-2 1xl:py-[0.55rem] 3xl:py-[0.7rem] mb-px 1xl:mb-1 3xl:mb-2 cursor-pointer transition-[background-color,border-color,padding,margin] duration-200 ease-in-out ${
+  const sharedClassName = `flex relative rounded-md items-center py-2 1xl:py-[0.55rem] 3xl:py-[0.65rem] mb-px 1xl:mb-1 3xl:mb-[0.3125rem] cursor-pointer transition-[background-color,border-color,padding,margin] duration-200 ease-in-out ${
     isCollapsed ? "mx-3 pl-[0.65rem]" : "w-full pl-4 pr-2"
   } ${
     isAnySubmenuActive
@@ -301,7 +301,7 @@ export const MenuItemWithSubmenu = ({
         )}
       </Tooltip>
       {!isCollapsed && isExpanded && (isSideMenuOpen || !isDesktop) && (
-        <div className="ml-[1.6rem] relative mt-[0.15rem] 1xl:-mt-[0.1rem] 3xl:-mt-[0.3rem]">
+        <div className="ml-[1.6rem] relative mt-[0.15rem] 1xl:-mt-[0.1rem] 3xl:-mt-[0.05rem]">
           <div
             ref={verticalLineRef}
             className="absolute left-0 top-0 w-0.5 bg-submenuTreeLine"
@@ -323,12 +323,12 @@ export const MenuItemWithSubmenu = ({
                 ref={(el) => {
                   submenuRefs.current[index] = el;
                 }}
-                className="block mb-px 1xl:mb-1 3xl:mb-2 -ml-[1.6rem] w-[calc(100%+1.6rem)] rounded-md relative focus-visible:outline-offset-[-2px]"
+                className="block mb-px 1xl:mb-1 3xl:mb-[0.2rem] -ml-[1.6rem] w-[calc(100%+1.6rem)] rounded-md relative focus-visible:outline-offset-[-2px]"
               >
                 <div className="absolute left-[calc(1.6rem+2px)] top-1/2 w-3 h-0.5 bg-submenuTreeLine"></div>
                 <div
                   onClick={handleMenuItemClick}
-                  className={`flex rounded-md items-center py-2 1xl:py-[0.55rem] 3xl:py-[0.6rem] pl-[3.2rem] w-full pr-2 transition ${
+                  className={`flex rounded-md items-center py-2 1xl:py-[0.55rem] 3xl:py-[0.5rem] pl-[3.2rem] w-full pr-2 transition ${
                     isActive
                       ? "bg-navItemActiveBg hover:bg-navItemActiveBgHover"
                       : "bg-navItemBg hover:bg-navItemBgHover"